requirements.txt
# 核心依赖
requests>=2.25.0
pandas>=1.3.0
numpy>=1.21.0

# 可选依赖（浏览器自动化）
playwright>=1.20.0

# 开发依赖
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800

# 数据处理
openpyxl>=3.0.0
xlrd>=2.0.0

# 日志和配置
pyyaml>=5.4.0
python-dotenv>=0.19.0

# 进度条
tqdm>=4.60.0

# 图像处理（可选）
Pillow>=8.0.0

# 网络请求增强
aiohttp>=3.7.0
asyncio-throttle>=1.0.0

# 异步数据库
aiomysql>=0.1.0

# 数据验证
pydantic>=1.8.0

# 时间处理
python-dateutil>=2.8.0

# 文件操作
pathlib2>=2.3.0

# 并发处理
concurrent-futures>=3.1.0

# 系统信息
psutil>=5.8.0
